requirements.txt
rich>=13.0.0
flask>=3.0.0
pyyaml>=6.0
pytest>=7.0.0
pandas>=2.0.0
pyarrow>=14.0.0
streamlit>=1.35.0
plotly>=5.20.0
